# apps/api/docs/openapi_v1.0.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CollegeERPAPI",
    "version": "1.0.0",
    "description": "\nCollege ERP System API\n\n**Generated:** 2026-02-03T12:39:45.480081\n\n**Locked Domains:**\n- Admission (v1.0.0) - Contracts locked with strict validation\n\n**Status:**\n- \u2705 Admission: Ready for frontend integration\n- \u23f3 Finance: In progress\n- \u23f3 Academic: In progress\n- \u23f3 Student: In progress\n\n**Breaking Changes Policy:**\n- Enum additions: Safe (backward compatible)\n- New optional fields: Safe\n- Required field changes: Requires migration + deprecation notice\n- Enum removals: 6-month deprecation period\n- Field type changes: Major version bump\n        "
  },
  "paths": {
...
    "/api/v1/campus/inventory/inventory/assets": {
      "post": {
        "tags": [
          "campus",
          "Campus - Inventory",
          "Inventory & Assets"
        ],
        "summary": "Create Asset",
        "description": "Add a new asset to the institution's inventory",
        "operationId": "create_asset_api_v1_campus_inventory_inventory_assets_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "campus",
          "Campus - Inventory",
          "Inventory & Assets"
        ],
        "summary": "List Assets",
        "description": "List assets with category filtering and search",
        "operationId": "list_assets_api_v1_campus_inventory_inventory_assets_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssetCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  },
                  "title": "Response List Assets Api V1 Campus Inventory Inventory Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/inventory/inventory/allocate": {
      "post": {
        "tags": [
          "campus",
          "Campus - Inventory",
          "Inventory & Assets"
        ],
        "summary": "Allocate Asset",
        "description": "Issue an asset to a person or department",
        "operationId": "allocate_asset_api_v1_campus_inventory_inventory_allocate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetAllocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetAllocation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/inventory/inventory/return/{allocation_id}": {
      "put": {
        "tags": [
          "campus",
          "Campus - Inventory",
          "Inventory & Assets"
        ],
        "summary": "Return Asset",
        "description": "Record the return of an issued asset",
        "operationId": "return_asset_api_v1_campus_inventory_inventory_return__allocation_id__put",
        "parameters": [
          {
            "name": "allocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Allocation Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllocationStatus",
              "default": "RETURNED"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetAllocation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/inventory/inventory/audit": {
      "post": {
        "tags": [
          "campus",
          "Campus - Inventory",
          "Inventory & Assets"
        ],
        "summary": "Perform Audit",
        "description": "Perform a physical stock audit for an asset",
        "operationId": "perform_audit_api_v1_campus_inventory_inventory_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetAuditCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetAudit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/inventory/inventory/uniforms": {
      "post": {
        "tags": [
          "campus",
          "Campus - Inventory",
          "Inventory & Assets"
        ],
        "summary": "Issue Uniform",
        "description": "Record uniform issuance to a student",
        "operationId": "issue_uniform_api_v1_campus_inventory_inventory_uniforms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniformAllocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniformAllocation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AllocationStatus": {
        "type": "string",
        "enum": [
          "ISSUED",
          "RETURNED",
          "DAMAGED",
          "LOST"
        ],
        "title": "AllocationStatus"
      },
...
      "Asset": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/AssetCategory"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "qr_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Code"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "PCS"
          },
          "total_stock": {
            "type": "integer",
            "title": "Total Stock",
            "default": 0
          },
          "available_stock": {
            "type": "integer",
            "title": "Available Stock",
            "default": 0
          },
          "reorder_level": {
            "type": "integer",
            "title": "Reorder Level",
            "default": 5
          },
          "unit_price": {
            "type": "string",
            "title": "Unit Price",
            "default": "0.00"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category"
        ],
        "title": "Asset",
        "description": "Main inventory and asset model - Physical Identity Owner"
      },
      "AssetAllocation": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "faculty_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Id"
          },
          "department_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Name"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          },
          "allocated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Allocated At"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "returned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned At"
          },
          "status": {
            "$ref": "#/components/schemas/AllocationStatus",
            "default": "ISSUED"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "asset_id"
        ],
        "title": "AssetAllocation",
        "description": "Tracking asset allocation to students, staff, or departments"
      },
      "AssetAllocationCreate": {
        "properties": {
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "faculty_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Id"
          },
          "department_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Name"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "asset_id"
        ],
        "title": "AssetAllocationCreate"
      },
      "AssetAudit": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "audit_date": {
            "type": "string",
            "format": "date-time",
            "title": "Audit Date"
          },
          "expected_stock": {
            "type": "integer",
            "title": "Expected Stock"
          },
          "actual_stock": {
            "type": "integer",
            "title": "Actual Stock"
          },
          "variance": {
            "type": "integer",
            "title": "Variance"
          },
          "audited_by": {
            "type": "string",
            "title": "Audited By"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "expected_stock",
          "actual_stock",
          "variance",
          "audited_by"
        ],
        "title": "AssetAudit",
        "description": "Inventory audit records"
      },
      "AssetAuditCreate": {
        "properties": {
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "actual_stock": {
            "type": "integer",
            "title": "Actual Stock"
          },
          "audited_by": {
            "type": "string",
            "title": "Audited By"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "actual_stock",
          "audited_by"
        ],
        "title": "AssetAuditCreate"
      },
      "AssetCategory": {
        "type": "string",
        "enum": [
          "UNIFORM",
          "IT_EQUIPMENT",
          "LAB_EQUIPMENT",
          "FURNITURE",
          "STATIONERY",
          "OTHERS"
        ],
        "title": "AssetCategory"
      },
      "AssetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/AssetCategory"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "qr_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Code"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "PCS"
          },
          "total_stock": {
            "type": "integer",
            "title": "Total Stock"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Unit Price",
            "default": "0.00"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "total_stock"
        ],
        "title": "AssetCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UniformAllocation": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "size": {
            "$ref": "#/components/schemas/UniformSize"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          },
          "is_paid": {
            "type": "boolean",
            "title": "Is Paid",
            "default": false
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "PENDING"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "item_name",
          "size"
        ],
        "title": "UniformAllocation",
        "description": "Specialized allocation for student uniforms"
      },
      "UniformAllocationCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "size": {
            "$ref": "#/components/schemas/UniformSize"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "item_name",
          "size"
        ],
        "title": "UniformAllocationCreate"
      },
      "UniformSize": {
        "type": "string",
        "enum": [
          "XS",
          "S",
          "M",
          "L",
          "XL",
          "XXL",
          "CUSTOM"
        ],
        "title": "UniformSize"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}